requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
celery-with-redis==3.0
diff-match-patch==20121119
Django==1.11.3
django-celery==3.2.2
django-guardian==1.4.9
django-haystack==2.8.0
django-import-export==1.0.1
django-jet==1.0.7
django-redis==4.9.0
django-tinymce==2.7.0
et-xmlfile==1.0.1
itsdangerous==0.24
jdcal==1.4
jieba==0.39
kombu==3.0.37
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.4.9
Pillow==5.0.0
pycryptodomex==3.5.1
python-alipay-sdk==1.7.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==3.3.0
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
